Remove old implementations of TabController (#1555)
