+++ generatedTypes/src/components/tabController/TabBarContext.d.ts
@@ -1,3 +1,17 @@
 import React from 'react';
-declare const TabBarContext: React.Context<{}>;
+import Reanimated from 'react-native-reanimated';
+interface TabControllerContext {
+    initialIndex?: number;
+    selectedIndex?: number;
+    items?: any[];
+    asCarousel?: boolean;
+    containerWidth: number;
+    pageWidth: number;
+    /** static page index */
+    currentPage: Reanimated.SharedValue<number>;
+    /** transition page index (can be a fraction when transitioning between pages) */
+    targetPage: Reanimated.SharedValue<number>;
+    carouselOffset: Reanimated.SharedValue<number>;
+}
+declare const TabBarContext: React.Context<TabControllerContext>;
 export default TabBarContext;
