+++ generatedTypes/src/components/tabController/index.d.ts
@@ -1,9 +1,5 @@
-import React, { Component } from 'react';
-import _ from 'lodash';
-import TabBar from './TabBar';
-import TabBarItem, { TabControllerItemProps } from './TabBarItem';
-import TabPage from './TabPage';
-import PageCarousel from './PageCarousel';
+import React, { PropsWithChildren } from 'react';
+import { TabControllerItemProps } from './TabBarItem';
 export { TabControllerItemProps };
 export interface TabControllerProps {
     /**
@@ -13,11 +9,15 @@ export interface TabControllerProps {
     /**
      * Initial selected index
      */
-    selectedIndex: number;
+    initialIndex?: number;
+    /**
+     * DEPRECATED: use initialIndex instead
+     */
+    selectedIndex?: number;
     /**
      * callback for when index has change (will not be called on ignored items)
      */
-    onChangeIndex?: (index: number) => void;
+    onChangeIndex?: (index: number, prevIndex: number | null) => void;
     /**
      * When using TabController.PageCarousel this should be turned on
      */
@@ -27,49 +27,21 @@ export interface TabControllerProps {
      */
     carouselPageWidth?: number;
 }
-interface StateProps {
-    selectedIndex: number;
-    asCarousel?: boolean;
-    pageWidth: number;
-    items: TabControllerProps['items'];
-    itemStates: any[];
-    ignoredItems: any[];
-    targetPage: any;
-    currentPage: any;
-    carouselOffset: any;
-    containerWidth: any;
-    registerTabItems: (tabItemsCount: number, ignoredItems: StateProps['ignoredItems']) => void;
-    onChangeIndex?: (index: number) => void;
-}
 /**
  * @description: A performant solution for a tab controller with lazy load mechanism
  * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/TabControllerScreen/index.tsx
  * @notes: This component is based on react-native-gesture-handler
  * @important: On Android, if using react-native-navigation, make sure to wrap your screen with gestureHandlerRootHOC
  * @importantLink: https://kmagiera.github.io/react-native-gesture-handler/docs/getting-started.html#with-wix-react-native-navigation-https-githubcom-wix-react-native-navigation
- * @gif: https://github.com/wix/react-native-ui-lib/blob/master/demo/showcase/TabController/Default.gif?raw=true, https://github.com/wix/react-native-ui-lib/blob/master/demo/showcase/TabController/PageCarousel.gif?raw=true, https://github.com/wix/react-native-ui-lib/blob/master/demo/showcase/TabController/CenterSelected.gif?raw=true
  */
-declare class TabController extends Component<TabControllerProps, StateProps> {
-    static displayName: string;
-    static contextType: React.Context<{}>;
-    static TabBar: typeof TabBar;
-    static TabBarItem: typeof TabBarItem;
-    static TabPage: typeof TabPage;
-    static PageCarousel: typeof PageCarousel;
-    static defaultProps: {
-        selectedIndex: number;
-        activeOpacity: number;
-    };
-    constructor(props: TabControllerProps);
-    static getDerivedStateFromProps(nextProps: TabControllerProps, prevState: StateProps): {
-        pageWidth: number;
-    } | null;
-    componentDidUpdate(_prevProps: TabControllerProps, prevState: StateProps): void;
-    get pageWidth(): number;
-    registerTabItems: (tabItemsCount: number, ignoredItems: StateProps['ignoredItems']) => void;
-    onPageChange: ([index]: readonly number[]) => void;
-    renderCodeBlock: (() => JSX.Element) & _.MemoizedFunction;
-    render(): JSX.Element;
+declare function TabController({ initialIndex, selectedIndex, asCarousel, items, onChangeIndex, carouselPageWidth, children }: PropsWithChildren<TabControllerProps>): JSX.Element;
+declare namespace TabController {
+    var TabBar: React.ComponentClass<import("./TabBar").TabControllerBarProps & {
+        useCustomTheme?: boolean | undefined;
+    }, any>;
+    var TabBarItem: typeof import("./TabBarItem").default;
+    var TabPage: typeof import("./TabPage").default;
+    var PageCarousel: typeof import("./PageCarousel").default;
 }
 declare const _default: React.ComponentClass<TabControllerProps & {
     useCustomTheme?: boolean | undefined;
