--- generatedTypes/src/components/tabController2/PageCarousel.d.ts
@@ -1,10 +1,0 @@
-/// <reference types="react" />
-/**
- * @description: TabController's Page Carousel
- * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/TabControllerScreen/index.tsx
- * @notes: You must pass `asCarousel` flag to TabController and render your TabPages inside a PageCarousel
- */
-declare function PageCarousel({ ...props }: {
-    [x: string]: any;
-}): JSX.Element;
-export default PageCarousel;
