+++ generatedTypes/src/incubator/index.d.ts
@@ -1,5 +1,4 @@
 export { default as ExpandableOverlay } from './expandableOverlay';
-export { default as TabController } from './TabController';
 export { default as TextField, TextFieldProps, FieldContextType } from './TextField';
 export { default as TouchableOpacity, TouchableOpacityProps } from './TouchableOpacity';
 export { default as TouchableOpacity2 } from './TouchableOpacity2';
